+++ app/classes/helpers.py
@@ -12,7 +12,7 @@ from datetime import datetime
 from OpenSSL import crypto, SSL
 from socket import gethostname
 import time
-from pprint import pprint
+#from pprint import pprint
 
 from app.classes.console import Console
 from argon2 import PasswordHasher